requirements.txt
# CIN Extraction and Plate Detection shared requirements
grpcio==1.54.0
grpcio-tools==1.54.0
numpy>=1.24.0
opencv-python>=4.8.0
easyocr>=1.7.0
openvino>=2023.0.0
pyyaml>=6.0
protobuf>=4.22.0
matplotlib>=3.7.0
Pillow>=10.0.0
scikit-image>=0.19.3
torch>=2.0.0  # Required by EasyOCR
torchvision>=0.15.0  # Required by EasyOCR
pillow>=10.0.0
matplotlib>=3.7.0
torchvision>=0.15.0
torch>=2.0.0
